[ET-VK] Integrate axis mapping into optimized matrix multiplication shaders + massive code cleanup
## Context

As title, this diff integrates `axis_mapping` into the optimized matrix multiplication shaders. As part of this change, I also decided to clean up the code as well:

* merge 2D and 3D cases into a single function. The behaviour for 2D/3D is controlled by `#ifdef` blocks. Some may find this slightly harder to read but I think it will be more maintainable in the long run since we won't need 2 copies of the core logic.

* Merge `addmm` and `mm` cases into a single shader. The only difference between the two is that `addmm` accounts for bias whereas `mm` doesn't. Previously we had separate shaders because the layout declarations would get kind of messy since the binding slot had to be declared explicitly, but the introduction of the `B` codegen variables takes care of that.

* General readability improvements.

I apologize in advance for the poor diff organization. Unfortunately I couldn't find an easy way to split the diffs given the magnitude of the changes.

fb:

## Performance Impact

An important consideration of the `axis_mapping` work is the latency overhead introduced by the extra logic needed to account of axis mapping when constructing texture positions. Empirically, I think the expected **maximum expected overhead is between 5 to 10%**.

However, it seems that latency does not always degrade. Suprisingly, **introducing axis mapping actually improves the latency of the optimized linear shader**. Of course, there were many other changes introduced as well, but during development I would try to isolate the latency impact of specific changes and it appears using axis mapping is indeed responsible for overall latency imprvoement. The actual mechanism by which accounting for axis mapping improves latency is not obvious to me, because the exact method of accounting for axis mapping changs the overall latency impact. For example, different techniques such as tracking a persistent `ivec3 pos` variable compared to constructing an `ivec3 pos` variable every loop iteration would yield drastically different results in different scenarios.

In any case, the benchmark results for encoder are:

| commit      | max (ms) | median (ms) | mean (ms) |
|-------------|----------|-------------|-----------|
| master      | 197      | 137         | 137       |
| this commit | 189      | 131         | 131       |


On the same device, the matmul operator benchmark was performed. The benchmark tests a M=32 x K =1024 x N = 512 matrix multiplication. The results are:

| commit      | matmul_optimized_tile_row_4_float median ms | linear_optimized_tile_row_4_float median ms |
|-------------|---------------------------------------------|---------------------------------------------|
| master      | 4.08898                                     | 3.48135                                     |
| this commit | 4.34442                                     | 3.59511                                     |

Evidently, evem the regular matrix multiplication shader was improved

Differential Revision: [D62444315](https://our.internmc.facebook.com/intern/diff/D62444315/)

ghstack-source-id: 241813259
Pull Request resolved: https://github.com/pytorch/executorch/pull/5222