[ET-VK] Remove unnecessary member structs from `Allocation` struct to reduce `vTensor` size
Pull Request resolved: https://github.com/pytorch/executorch/pull/5392

## Context

Recently, the size of `vTensor` has grown dramatically, due to adding multiple metadata UBO member variables to store Tensor metadata. The size of the `vTensor` object was over 2KB on 64-bit machines.

One large contributor to the size of `vTensor` was the `Allocation` class, which was storing various structs as class members. As a result of keeping these structs in memory, the size of each instance of `Allocation` was over 200 B.

These structs are not necessary to keep in memory, thus they can be removed. Once removed, the size of `vTensor` is now halved to be 1KB.

The size of `vTensor` can be further reduced by sharing metadata UBO member variables and being more conservative with keeping tensor metadata in memory.
ghstack-source-id: 242810657
@exported-using-ghexport

Differential Revision: [D62660980](https://our.internmc.facebook.com/intern/diff/D62660980/)