[ET-VK][BE] vTensor cleanup 5/N - clean up `indexing_utils.h` and clarify function names
Pull Request resolved: https://github.com/pytorch/executorch/pull/5443

## Context

The goal of this diff is to clean up the `indexing_utils.h` header file to introduce more consistent terminology for tensor properties and improve the function names so that they are crystal clear on what they are doing.

## Notes for reviewers

As with the last diff, the majority of meaningful changes are in `indexing_utils.h` and the rest are changes to function names. There should be no functionality changes, with the exception of re-writing functions which interact with texture position so that input arguments can be const.
ghstack-source-id: 243365712

Differential Revision: [D62901892](https://our.internmc.facebook.com/intern/diff/D62901892/)