[ET-VK] Update `RemoveLocalScalarDenseOpsTransform` to tag scalar tensors as well
## Context

See the new docstrings added to `remove_local_scalar_dense_ops` for more details on what the pass is trying to achieve.

The goal is to mark tensors that are consumed as scalars via `tensor[0].item()` as "scalar tensors" that will be represented as a `SymInt` object in the vulkan delegate instead of a regular `Tensor` object.

This diff also adds an `__init__.py` file to the `_passes` folder to make it easier to include Vulkan passes from one place.

Differential Revision: [D64139867](https://our.internmc.facebook.com/intern/diff/D64139867/)

ghstack-source-id: 247163956
Pull Request resolved: https://github.com/pytorch/executorch/pull/6069