[ET-VK][ez] Apply rotary embedding as Module (#6391)
* [ET-VK][ez] Apply rotary embedding as Module

## Context

As title. Wrap the `apply_rotary_emb` function call in a `nn.Module` to make it easy to perform a source module replacement for rotary embedding calculation.

The Vulkan delegate will use the source module replacement technique to insert a custom op to calculate rotary embeddings.

Differential Revision: [D64697589](https://our.internmc.facebook.com/intern/diff/D64697589/)

[ghstack-poisoned]

* Update on "[ET-VK][ez] Apply rotary embedding as Module"


## Context

As title. Wrap the `apply_rotary_emb` function call in a `nn.Module` to make it easy to perform a source module replacement for rotary embedding calculation.

The Vulkan delegate will use the source module replacement technique to insert a custom op to calculate rotary embeddings.

Differential Revision: [D64697589](https://our.internmc.facebook.com/intern/diff/D64697589/)

[ghstack-poisoned]