[llama-mm] Add export-friendly tile position embedding
Summary:

Before we make a decision on whether torchtune takes this
export-friendly version of `TilePositionEmbedding`, we put it under
`extension/llm` so that users can start to use it.

Added unit tests to make sure the behavior is the same as the reference
implementation in torchtune and export/AOTI/ET all working properly.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 96e7da25ae81395b9a37f0382b2222e3443273f8
Pull Request resolved: https://github.com/pytorch/executorch/pull/6650